Split SSH server into connection, command, and LFS auth files
server.go (617 lines) mixed connection/session plumbing with git command
serving and git-lfs-authenticate. Move exec parsing and git serving to
command.go (with the mirror-access check extracted) and the LFS auth response
to lfs_authenticate.go. Pure moves otherwise.
